+++ frontends/web/src/routes/bitsurance/bitsurance.tsx
@@ -13,6 +13,7 @@ import { Column, ColumnButtons, ResponsiveGrid } from '@/components/layout';
 import { View, ViewContent } from '@/components/view/view';
 import { useDarkmode } from '@/hooks/darkmode';
 import { i18n } from '@/i18n/i18n';
+import { BitsuranceLayout } from './bitsurance-layout';
 import style from './bitsurance.module.css';
 
 type TProps = {
@@ -83,59 +84,64 @@ export const Bitsurance = ({ accounts, code }: TProps) => {
   }
 
   return (
-    <View fullscreen={false}>
-      <ViewContent>
-        <p className={style.noVspace}>{t('bitsurance.intro.text1', { amount })}</p>
-        <div className={style.gridContainer}>
-          <ResponsiveGrid col="2" textAlign="start">
-            <Column asCard>
-              <h3 className={style.title}>
-                {t('bitsurance.insure.title')}
-              </h3>
-              <p className={style.cardBody}>{t('bitsurance.insure.text')}</p>
-              <ul className={style.clean}>
-                <li><Checked/><span>{t('bitsurance.insure.listItem1')}</span></li>
-                <li><Checked/><span>{t('bitsurance.insure.listItem2')}</span></li>
-                <li><Checked/><span>{t('bitsurance.insure.listItem3')}</span></li>
-              </ul>
-              <p className={style.cardBody2}>
-                {t('bitsurance.insure.text2')} {' '}
-                <A href={getBitsurancePageLink()}>{t('bitsurance.intro.link')}</A>.
-              </p>
-              <p className={style.cardBody2}>
-                {t('bitsurance.insure.text3')}
-              </p>
-              <ColumnButtons className={style.ctaButton}>
-                <Button onClick={maybeProceed} primary>
-                  {t('bitsurance.insure.button')}
-                </Button>
-              </ColumnButtons>
-            </Column>
-            <Column asCard>
-              <h3 className={style.title}>
-                {t('bitsurance.detect.title')}
-              </h3>
-              <p className={style.cardBody}>{t('bitsurance.detect.text')}</p>
-              {!insuredAccounts.length && scanDone && (
-                <p className={`${style.cardBody2 || ''} ${style.errorMessage || ''}`}>
-                  {t('bitsurance.detect.notInsured')}
+    <BitsuranceLayout
+      accounts={accounts}
+      code={code}
+    >
+      <View fullscreen={false}>
+        <ViewContent>
+          <p className={style.noVspace}>{t('bitsurance.intro.text1', { amount })}</p>
+          <div className={style.gridContainer}>
+            <ResponsiveGrid col="2" textAlign="start">
+              <Column asCard>
+                <h3 className={style.title}>
+                  {t('bitsurance.insure.title')}
+                </h3>
+                <p className={style.cardBody}>{t('bitsurance.insure.text')}</p>
+                <ul className={style.clean}>
+                  <li><Checked/><span>{t('bitsurance.insure.listItem1')}</span></li>
+                  <li><Checked/><span>{t('bitsurance.insure.listItem2')}</span></li>
+                  <li><Checked/><span>{t('bitsurance.insure.listItem3')}</span></li>
+                </ul>
+                <p className={style.cardBody2}>
+                  {t('bitsurance.insure.text2')} {' '}
+                  <A href={getBitsurancePageLink()}>{t('bitsurance.intro.link')}</A>.
                 </p>
-              )}
-              <ColumnButtons className={style.ctaButton}>
-                <Button
-                  onClick={() => detect(true)}
-                  disabled={scanLoading}
-                  secondary
-                >
-                  {isDarkMode ? <SyncLight/> : <Sync/>}
-                  {t('bitsurance.detect.button')}
-                </Button>
-              </ColumnButtons>
+                <p className={style.cardBody2}>
+                  {t('bitsurance.insure.text3')}
+                </p>
+                <ColumnButtons className={style.ctaButton}>
+                  <Button onClick={maybeProceed} primary>
+                    {t('bitsurance.insure.button')}
+                  </Button>
+                </ColumnButtons>
+              </Column>
+              <Column asCard>
+                <h3 className={style.title}>
+                  {t('bitsurance.detect.title')}
+                </h3>
+                <p className={style.cardBody}>{t('bitsurance.detect.text')}</p>
+                {!insuredAccounts.length && scanDone && (
+                  <p className={`${style.cardBody2 || ''} ${style.errorMessage || ''}`}>
+                    {t('bitsurance.detect.notInsured')}
+                  </p>
+                )}
+                <ColumnButtons className={style.ctaButton}>
+                  <Button
+                    onClick={() => detect(true)}
+                    disabled={scanLoading}
+                    secondary
+                  >
+                    {isDarkMode ? <SyncLight/> : <Sync/>}
+                    {t('bitsurance.detect.button')}
+                  </Button>
+                </ColumnButtons>
 
-            </Column>
-          </ResponsiveGrid>
-        </div>
-      </ViewContent>
-    </View>
+              </Column>
+            </ResponsiveGrid>
+          </div>
+        </ViewContent>
+      </View>
+    </BitsuranceLayout>
   );
 };
